First implementation of the CADO lattice siever which outputs relations!
Some parameters are still hardcoded, thus this is not yet ready for production.
On the rational side, since we have no sieving yet, I have implemented 
Bernstein's idea with a product tree of factor base primes. It is not very
fast, but the advantage is that it completely splits the norms into the
factor base primes and the large primes, thus it makes trial division much
easier.

On our RSA155 running example, with q=16777291, rho=4078255, I=2^13, J=2^12,
it outputs 10 relations in about 56 seconds of a 2.4Ghz Opteron, and it
remains 8 after checknorms. Half of the time is spent reading the factor base,
thus the true time is about 28 seconds.
In comparison, Franke/Kleinjung's implementation takes about 2.8s for the
same parameters, and outputs 14 relations (we loose 6 relations because of
special_case_0 and special_case_p I guess).

Now we have a running code, we can identify the slow phases (in particular the
estimation of the norm of F(a,b) is a first candidate), and improve them.


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@828 de4796e2-0b1e-0410-91b4-86bc23549317