Some improvements in the lattice siever: * a global norm bound is now computed independently of the given special-q   (thus the factor base is read and scaled only once) * it can now loop over several special-q's
What remains to be done:
* complete rewrite of the norm initialization (the current routine takes more
  than half of the total time!)
* several possible speedups in Bernstein's "gcd tree" algorithm on the rational
  side, and compare with sieving on the rational side
* deal with special_case_0 and special_case_p (will give some reports which
  are currently missed)
* and of course many bugs to fix...

Anyway you can start playing with the lattice siever, at your own risk:

$ ./las -poly ... -fb ... -q0 a -q1 b > rels_raw

will search relations for all special-q in [a, b[ (-poly and -fb take the
same files as the line siever).

Just don't perform major changes in the code, since it is in heavy development.


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@840 de4796e2-0b1e-0410-91b4-86bc23549317