To use the new features correctly.
To be clearer: when resume is activated in the mpi version [via merge.sh],
then a merge of level 2 is performed ON THE MASTER ONLY. Then the merge is
resumed on all procs as before. The master works on a smaller window of the
matrix to decrease the memory used, so that it can fit in 8 Gb for c155.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@1233 de4796e2-0b1e-0410-91b4-86bc23549317