Let's say Markowitz can be tried. To use it, you have to uncomment the ifdef in merge_opts.h . Then
cd linalg; make clean; make

This is not the default right now, since we need to experiment.

Features:
* ifdef so that swar and markowitz cannot live at the same time (easier for
cleaning / development); can be changed if really needed; code easier
to follow than swar, probably; the input/output, etc. is common to both
swar and markowitz.
* Same parameters as for swar; same strategies with -forbw.
* tested for c59, c80, c95, rsa100.
* frankly, this is just another solution of the same optimization pb, so that
comparing on the same input is not enough. Experiments are needed.

If you want to limit c/N as in Papadopolous's slides, use

	-forbw 3 -coverNmax 70

but again, this depends on bw...

Have fun. I will continue testing for larger matrices at the same time, since
some strategies need to be tested again.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@1790 de4796e2-0b1e-0410-91b4-86bc23549317