I seem to have written the postprocessing stuff. Final result shorter than I once thought. Modifications were needed in replay.c, adding a new option.
For instance for rsa100, I found
abondance% grep We Examples/rsa100/rsa100.10x41x101/log
# Weight(M_small) = 14701383

after postprocessing:
# Weight(M_small) = 14699843

so we gain something. Almost nothing is gained on the c80.

Still experimental and files + Makefile need some cleaning, after some more
tests. Example of use is in merge_linalg_sqrt2.sh .

If you want to use it:

1. change true to false in merge_linalg_sqrt2.sh (look for postprocess).
2. compile linalg with USE_MARKOWITZ on (sorry for this, it should work
with swar, but this is not the case for organizational reasons).
3. cd linalg; make postprocess (it is not connected in Makefile due to 2.
Do not change this yet, please.).

For the time being, two small matrices are built. One would be enough, but
it is more convenient for comparisons.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@1833 de4796e2-0b1e-0410-91b4-86bc23549317