[cantor/gm] put an end to cantor/gm stuff.
Some dirty tricks are disabled so that it safely compiles on 32-bits
also.  There's yet a fully truncating version to write, but its running
time would be similar to the truncated cantor in any case. Of course
interpolation isn't there either, for the same reason. Time to put
this to sleep until avx comes out.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@1863 de4796e2-0b1e-0410-91b4-86bc23549317