[bwc] enable (and fix) advanced barrier waits.
This has the advantage of fixing the nasty thread_agreement bug, at a
moderate cost.  Furthermore, it exerts some real-life testing on code
which is capable of emulating pthread_barrier_wait on platforms that
don't have it. Thus it's now configured that way, and threads should work
on the mac.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@2100 de4796e2-0b1e-0410-91b4-86bc23549317