Using Paul's product-of-all-Z-coordinates trick instead of batch inversion to canonicalize precomputed points, saves an extgcd.
Initialising accumulator to Pj_z[0] (product of the Z-coordinates
of all precomputed points that are not the end-of-stage-1 point)
to avoid missing factors wrt. the batch inversion.

Fixed stupid bug: comments in ellM_add() were wrong, and based on
them the test for identical input points was also wrong.

Using mod_sqr() where possible.


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@2171 de4796e2-0b1e-0410-91b4-86bc23549317