[bwc] added an extra flag to bwc.pl, for mpi usage.
We should be able to routine-check mpi setups now.

(the following doc, for once, goes not only in the commit log but also in
the bwc.pl file itself).

The mpi_extra_args argument is used to pass information to the mpiexec
command. The idea is that mpiexec, the mpi driver program, may need
additional info to properly setup communications between jobs. As an
example, on Fedora machines with openmpi:

./build/x86_64/linalg/bwc/bwc.pl :complete matrix=/net/tiramisu/localdisk/thome/mats/c90b wdir=/local/rsa768/tmp/c59 mn=64 mpi=2x1 thr=2x4 hosts=patate,tiramisu interval=10 mpi_extra_args='--mca btl_tcp_if_exclude lo,virbr0'

This tells mpi not to try routing traffic through either the lo or the
virbr0 interface. For the former, it's already openmpi's default
behaviour (unless /etc/hosts is screwed up, which this file does tend to
be sometimes). For virbr0, it's again some local mess, but each machine
having this class C network defined, openmpi can't really tell whether
they're connected or not -- and of course they're not.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado-nfs/trunk@91 3eaf19af-ecc0-40f6-b43f-672aa0c71c71