ugly, ugly, ugly. ugly.
This patch allows fine-grain selection of whether some given program is
an mpi program or not. cmake does not allow per-target selection of the
compiler, which would be the natural way to go. So...

If ever this breaks, this would have to be reverted, as we could live
without this feature (although neat from the flexibility point of view).

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado-nfs/trunk@102 3eaf19af-ecc0-40f6-b43f-672aa0c71c71