added -Wno-unused-result.
This warning is important only for gcc-4.5 because on some occasions we
are not checking the return value of fscanf/asprintf/fgets/read/write/etc
(gcc becomes pesky about this). For gcc-4.4, adding this option
causes innocuous warnings _only if_ other warnings are already present,
so it is harmless. The fine points of the rationale are at:

http://gcc.gnu.org/bugzilla/show_bug.cgi?id=28322

Now if someone still uses gcc-4.3 or older, then we're in trouble, and
need to add some cmake mechanism to check whether the flag is supported.
Indeed, older gcc's will error out.

tiramisu /tmp $ gcc34 -W -Wall -Wno-unused-result a.c
cc1: error: unrecognized command line option "-Wno-unused-result"

So -- not sure it's an urgent matter, but if someone has cmake skills to
do guard this addition, that would be cool.

Other option: fix the 40+ such warnings we get. Most are completely
irrelevant (who cares about a failing asprintf in cado-nfs, really ?).

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado-nfs/trunk@208 3eaf19af-ecc0-40f6-b43f-672aa0c71c71