[bwc] new feature for hu-u-u-ge matrices: load from remote
For working on grids, transferring the full matrix before being able to
dispatch it to clients makes the whole thing 1) boring (have to wait
hours), 2) sometimes not possible (/tmp/'s can be too small). Since the
new balancing code now works completely in streaming mode, it is now
possible to kill the middle man (the matrix file on the head node), and
content ourselves with the matrix being in some remotely accessible
location.  Here we use libcurl as a backend for fetching the matrix.
This is basically the same idea as doing a popen("wget -O- <uri>"),
albeit slightly less fragile, and more fun.

Also taken the occasion to split balancing.c into different logical
parts. It's half-baked, in that it would be possible to do more, but
that's good enough for now.

Tested on the rsa768 matrix.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado-nfs/trunk@277 3eaf19af-ecc0-40f6-b43f-672aa0c71c71