warning killing party for gcc-4.6.1 (tiramisu has it).
Alex, if you care to have a look: facul_doit.c contained some runtime
checks which were in fact compile-time checks. I've changed them to the
latter, since the never-taken branches were actually triggering warnings.

For example, facul_doit_ul checks if the modulus is larger than 96 bits
(which it can't, but gcc sees only a runtime check) -- if it happens to
be, then it calls modredc15ul_initmod_uls with an array which actually
isn't that large. Hence the warning, which surprises a bit...

/tmp/cado/sieve/ecm/facul_doit.c: In function ‘facul_doit_ul’:
/tmp/cado/utils/modredc_15ul.h:142:8: warning: array subscript is above array bounds [-Warray-bounds]

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado-nfs/trunk@937 3eaf19af-ecc0-40f6-b43f-672aa0c71c71