Propagate {DY,}LD_LIBRARY_PATH across ssh
This is very clearly *not* done the way it should. There's no reason the
path on the local host might have any meaning on the remote side. The
mach_desc file should include this kind of remote host information (it
does so for PATH already). This file should be properly created by
factor.sh according to needs (for testing, mostly), and more important,
read_machines in cadofct.pm should obey a `libdir'' variable there.