Reorganize the small fb
The idea is that now the fb are no longer split into ``nice'' and ``bad''
primes, but we keep markers saying which primes are bad. This is an
intermediary step towards splitting the stuff more aggressively.

Experiments seem to indicate that if there's a speed impact, it's at
worst very mild (0.5%).