cado.h must be on _top_ of compilation units!
The reason for this is that nobody knows which magic is used by Joe
User's favourite unix system to expose some given prototype from the
libc. This applies, in particular, to functions we should expect from
stdlib (imagine posix_memalign, for instance).

cado.h tries its best to activate the different flags for this purpose.

It is safe to include cado.h on top, because it promises to only define
preprocessor strings.