Payday: factor out small sieve init.
Now that the factor base ordering has the resieved primes in a contiguous
zone, it's very easy to factor out the small sieve initialization stuff,
and take the next_position (now ssdpos) fields completely out of the way
(they're local and changing, while the ssd are shared and constant for a
given special q).

It's not expected to be a big win, but it does clear up a large part of
the mess.