allocate space for las_report on the heap, not the stack.
Two arrays of 256*256 unsigned longs sum up to 1MB. And several are on
the stack frames of different functions in total. So it's bug prone
(ulimit -s might be low), annoying for valgrind (which uses heuristics to
decide that the process is ``switching stack'' if rsp moves fast), and
it's also conceivably a moderate performance hit, although I'd doubt it.