Propose improved mechanism for parallel remote_cmd
cadofactor can't deal with large clusters properly. Okay, the cluster
scripts are more adapted to this, but OTOH a simple illustrating example
_should_ work, and not take ages pinging jobs.

This mechanism offers some advantages. At the moment only the initial
probe uses it, but it eliminates the associated delay completely. It also
has a downside, is that it's limited by the OS maximum number of open
files (typically 1024). So more than 1024-epsilon sub-jobs are a no-go
here.