Yet another interface for the fppol_get/set_ui functions.
Added fppol*_(monic)?_get_ui_mul_ti(), which first shifts the polynomial
(i.e., multiplies it by a power of t) before the conversion.  This is
useful to compute the starting point of a line, i.e., a position
corresponding to (0,j) = j*t^I, without having to waste time converting
the first I coefficients.

Also, for monic polynomials, the leading group of coefficients is
always converted using the monic look-up table, regardless of the number
of terms. As a side effect, there is no need to specify the number of
terms anymore.