Improve the matmul_bucket_methods parameter
This parameter allows to select the various algorithms which are entitled
to contribute to the sparse matrix structure. Currently there are some
shortcomings in that very few setups are supported, and others silently
lead to bogus computations. This patch improves this. In particular we
intend to provide some ``only small blocks'' algorithms, for instance.

Note that matmul_bucket_methods is an internal obscure parameter, so this
change isn't too important.