Binary for testing ffspol irreducibility.
This is a Monte-Carlo algorithm: it reduces the ffspol for many small
irreducible ells. If at some point it finds that the reduction is
irreducible, then the original polynomial was also irreducible. Otherwise
we don't know, but we suspect that it is reducible.
(if the Galois group is generic, the probability that an irreducible
stays so after reduction is in 1/deg(f)).