This helps antique gcc's to cope with extern inline.
As per the documentation, a debug build need not actually inline
functions. However, it should abide by the law for extern inline, which
is to avoid emitting an addressable code for such functions. So there's a
potential to not have the function at all. cado-nfs so far is relying
on the compiler/linker being in good cooperation, and emit e.g. weak
symbols for such functions. gcc-4.2.1 on openbsd 5.2 does not do so, and
fails the build. This corrects the issue, and does not seem to cause harm
on other platforms (admittedly, it's not pretty).