Got rid of bogus random_generation.[ch] header.
This code was used only to provide system-independent random behaviour.
But even so, it was half-baked, not enabled by default, and required
changes for compilation.

If I ever want to use such a feature, providing an LD_PRELOAD is probably
cleaner, and less intrusive (the only case where I see it's really useful is
prep in bwc).