Reorganized las in view of the descent.
The aims are twofold:
 - allow the special-q's to be given in the form of a ``todo list''.
 - allow some crucial parameters of the siever configuration to be
   reactive to the special-q to be handled.

The most serious usage is for the descent. While this may also serve
normal usage for relation collection, it's mostly anecdotic.

The number of relations produced for given benches may vary slightly,
since the norm initialization is now made dependent only on the bit size
of the special-q, not on q itself. If we were to redo
sieve_info_init_norm_data once per sq, we could work around this newly
introduced inaccuracy.

Note that _yet_, we have not enabled support of larger special-q's at
this point.