Cleanups for plingen
This changes the entry points for the bw_lingen functions, and fixes a bug
when plingen was used in mpi mode for a below-mpi-threshold sequence.

Additionally, we now assert that polymat::size is set early, just as a
move towards policing the use of these structures. The idea is to move to
either lazy/not-allocated-yet or allocated-and-wont-realloc.  Eventually
this will make it easy to turn the point of view around between matrices
of polynomials and polynomial of matrices (note though that the
bwmat_copy_coeffs catch-all call must go as well).

Some unused code has been deleted.