improved some plumbing code in bwc
For mksol to deal with extracting only part of the solution space, and
for the whole thing to properly support n>1 for prime fields, we need to
expand the functionality a bit. This commit introduces a new obscure
command-line argument (nsolvecs) for bwc, to be used only by mksol and gather

For consistency with the binary case, the matrix returned by plingen is
now transposed.