Refactoring of fb_read() code
Trying to understand what's going on in an effort to make fb_read() split
the factor base into pieces for individual threads, so that we don't have
to modify the factor base again later. The purpose of that is to allow
mmap()-ing the already-split-up factor base, without having functions later
try to change the data.