Replay: allow index construction for DL.
Now, the index is constructed at the same time as the rows are merged.
This has the advantage of saving one pass over the data. On the other
hand, it requires more memory. Replay is not supposed to be critical in
any way, but is simplifies the code.

The so-called "optimize" phase of replay is desactivated.