Fixed bogus specializations
For some reasons the specializations were not called at all, which is
somewhat funny. In fact, the bulk of ::schedule() and ::complete() is
better kept identical between the two versions. The only thing is that
the level_processed[] array is handled differently.

I think it's possible to do better, and get completely rid of
level_processed[] in the fast case without too much code noise.