Added a command-line switch --force-posix-threads to all filter_rels utilities.
This applies only to the passes where there are no multithreaded steps in
the processing (i.e. *NOT* when dup2 reads new files, which is done with
posix threads unconditionally, because several threads compute the roots).

I might be mistaken, but tests on my machine let me think that the
benefit of using busy waits instead of pthreads is in fact much lower
that I once thought. It might be dependent on how the machine is loaded,
this sort of things.  Please try. If using posix threads is seen to never
degrade performance, we may drop the C++ syntax, and be satisfied with
only a single implementation.

(debug builds with -O0 are indeed much slower with posix threads, but
that's not a valid reason for preferring a complicated code to a simpler
one).