Improve random matrix generation, for full staging of LA runs.
We now offer the possibility to bypass the bottleneck of generating a
huge matrix and then dispatching it to many nodes, prior to getting an
idea of how expensive LA would be. In effect, this change makes the
secure, krylov, and mksol programs essentially autonomous for fake
matrices (bwc.pl does not understand this). Creation of local cache files
is delegated to the parallel jobs, which is the whole point behind. This
saves several hours of waiting before running tests.