Use param_list doc functions in bwc binaries.
This is not a "complete" patch in the sense that I'm still using
param_lists to pass around parameters to inner functions. So there's some
ugliness left. But I find it less ugly than being bothered with tons of
extra structs all around the place just to host the result of parsing
parameters, and before calling actual ctors for internal structures.