Fix for MPI builds.
split and acollect now also depend on MPI, which is a bit of a shame. On
the other hand, I'm pretty happy with having gotten rid of the bw-common
/ bw-common-mpi tandem, therefore so be it.

In the medium term, I'd like these two binaries to disappear anyway. So
fine-cutting the dependencies isn't that much of an important goal.