Got rid of build-everything-before-testing-anything misfeature.
I finally found how to do this with cmake hacking. Based on the second
(not best voted) answer there:

http://stackoverflow.com/questions/733475/cmake-ctest-make-test-doesnt-build-tests

A consequence is that both "make test" and "make check" now work.

Buildbots will be delighted to see that the number of tests soared by
almost a factor of two.

There could be some too-coarse-grain or missing dependencies floating
around. And the macros are arguably clumsy at times, e.g. names are not
good and so on.