add_full_cols parameter becomes lcideals and works for nb_polys != 2
lcideals parameter specifies that columns should be added in the renumbering
table to take into account the ideals corresponding to the leading coefficients
of polynomials. It is only needed for DL with non monic polynomial. The number
of additional columns is automatically computed by renumber.c (when called by
freerel) and the columns are set by dup2 (in print_relation).