Simplify cado-nfs-client daemonization.
The double-fork pattern is useless if the grandparent does exit()
straight away. At least so it seems.

(the previous approach with SIGUSR2 was doomed because it was not
checking the condition atomically after signal.pause(). signal.pause() is
not pthread_cond_wait. The latter has a mutex, for a reason!).