Bug fix. mksol computations with the git version prior to this were (possibly) wrong.
If you have a running mksol, it is advisable to kill it right now, and
let the checkpointing mechanism do the rest of the computation (with a
recompiled binary). Otherwise chances are that the very last chunk will
be incorrect (the files S.solsIII-JJJ.XXX-YYY.TTT with largest TTT
value).