- I modified the read_data function. Since we always compute a p-maximal order starting from the identity, it only reads the degree and the coefficients of the polynomial in the file you give in command line - I'm not sure if I get the right minimal polynomial, I'll try to compare with magma
For now, the program :
- Reads the degree and the coefficients of one polynomial in one file
- Computes the p maximal order Ok (p is still given in command line)
- It has an element theta of Ok (its expression is on the basis of generators of Ok, the ones printed by the program) ; but theta is directly written in the code, and for now, its written to work with a polynomial of degree 8
- The minimal_poly_of_mul_by_theta function computes the matrix of multiplication by theta, builds the (n,n^2) matrix, computes its kernel, and build theminimal polynomial of (x -> x*theta) with the Kernel.
However, I'm still not sure how to get the minimal polynomial.
I guess it's a matter of using the rows in the kernel to build one row with the most zeros (mod p) on the right.
But I noticed that the first row of the matrix of the kernel often seems to be the minimal polynomial. (I compared with magma).
I'm trying to see if there are example where it doesn't work.