minimal_polynomial_of_mul_by_theta doesn't work
Code compiles, there seems to be no memory leaks (although valgrind tells me about still reachable block, but I don't know if it's a problem or not).
But the function which returns the minimal polynomial of multiplication by one element in one order doesn't seem to work. (Compared with magma.