Building the list of generators of one ideal : not working.
I made a cxx_mpq_poly structure (simply containing cxx_mpz_poly and a mpz_t),
and wanted to make a vector of cxx_mpq_poly. Code compiles and is executed.
Each generator (one polynom) is represented before being pushed on gens, and after
but the denominator isn't the same. It's because the denominator is the mpz_t from cxx_mpq_poly, and it is freed after being pushed. Thus it contains something we don't know.
For now, I'm trying to correct that.