Valuation of one ideal on another : working.
How to enter the data ?
- in the usual file, type degree, coeff of polynomial, and coeff of ONE generator of ideal

What does the code do ?
- Same as before : compute the factorisation_of_prime(K,p), K computed from g, and p entered in command line
- print the ideals in the factorization of p*O (O being the p-maximal-order), and their multiplicities
- computes the ideal J  genereated by your generator (for some reason I don't know, magma doesn't return the same basis. But I tested : they're equivalent, they can generate each other)
- Test the valuation of J on the last ideal I of the list you got after the factorisation.

Why the last ideal ? Well, that was just for the test. We could test with other ideals.
And also : you can only choose one ideal J generated from one generator ; it's not done (yet) for two generators, and I don't see (for now) how you could give an ideal which is a product of other ideals.