Problem solved. Now maximal order should always work.
Explanation : it wasn't in the nullspace, but in the (n,n^2) matrix.
When I computed n coefficients (e.g. one polynomial) from one generator of O and of of I_p, I had to store those coefficients in a row matrix.
But I only filled the coefficients in the matrix from 0 to degree of the polynomial.
And I didn't flush the other coefficients left.

Thus, when the degree was too small, some coefficients were not overwritten, whereas they should have been.