Modify mksol so that we now run it only once per solution block.
This also goes with several sanitizing changes to various places of the
code. For the moment, interleaving is broken, and likewise for the wider
arithmetic types used within matmul (e.g. don't run with ys=0..128).

We can fairly safely assume that 32-bit builds are broken by this commit,
and I can't go check this right now.

Also:
    the Y files are gone.
    lingen produces generator into multiple independent chunks.
    the split binary is gone.
    unified bwc-trace and bwc-ptrace.
    some doc updated.