More stats. Basic idea seems ok. Also a minor threading change.
There's one gotcha though. We have #workspaces = #threads+1 (except when
we're single-threaded anyway). That seems to imply that using the thread
pool for handling the workspaces is a bad idea, because synchronization
will be expensive eventually.

Despite that, tests seem to indicate that there's no measurable loss so
far (actually, a 1.5% speed-up) for the hsnfs1024 use case (10
single-threaded las processes on each catrel node), and about similar for
the rsa220 use case (0.6% better, 2 8-core jobs on catrel). I did witness
a 1% loss or so on my desktop machine for a 4-core hsnfs1024 job, but for
less expensive tests.