recompute small sieve positions at each region, for the exceptional small primes.
Those are rare, so one would expect that the cost is negligible.

We also get rid of small_sieve_skip_stride, and do the corresponding
work directly within the sieve functions.

Validation: tested on catrel on an rsa220 sample, jobs 25284 to 25315.
The addition "small sieve + resieve gains about 1% on average. However
we seem to lose some reports. Investigation needed.