removed L2_combined_skewness (see rationale below)
Rationale: L2_combined_skewness tries to optimize the Murphy E-value by local
optimization. However, the Murphy E-value is estimated by simple sampling in
some *fixed* directions. Thus, the local optimization will likely adjust the
skewness so that one of those directions is close to a root of f (or g), which
will give a very high contribution to E which will perturb the estimation of E.