Introduce the (obscure) parameter maxwuerror.
maxwuerror is 2 by default, but can be increased if jobs get often
killed in a way that leads them to fail with something that is
understood as grave by the server, but is in fact rather innocuous.

I was tempted to document maxwuerror in the exact same place where
maxfailed and maxresubmit are documented. Alas, there does not seem to
be such a place...